fix(topology): bound SNMP index conversions and escape test array literals
intFromOIDSuffix parsed with Atoi and AssembleFdbEntries narrowed the
result to uint32, so an ifName row at index 4294967297 aliased ifIndex 1
and labelled it. Index suffixes now parse as Integer32 (ParseUint bitSize
31); uint32Value checks before converting, the VLAN id is bounded before
the uint16 conversion, and the LLDP address length compares without
narrowing (CodeQL go/incorrect-integer-conversion). The vertical
integration test escapes backslashes in its Postgres array literal
(js/incomplete-sanitization).

+++ agent/internal/discovery/adjacency_ports.go
@@ -350,7 +350,7 @@ func lldpManagementAddresses(pdus []gosnmp.SnmpPDU) map[string][]string {
 	sets := map[string]map[string]bool{}
 	for _, p := range pdus {
 		idx, ok := suffixParts(p.Name, snmppoll.LldpRemManAddrIfSubtypeOID)
-		if !ok || len(idx) < 5 || int(idx[4]) != len(idx)-5 {
+		if !ok || len(idx) < 5 || uint64(idx[4]) != uint64(len(idx)-5) {
 			continue
 		}
 		raw := make([]byte, 0, len(idx)-5)
